requirements.txt
faiss-gpu
sentence-transformers
pandas
numpy
pyarrow
fastparquet
torch
catboost
ipywidgets
scikit-learn
